isolate global namespace of plugins.
move away functions from plugins/__init__.py
