+++ nbodykit/plugins/HaloFilePainter.py
@@ -1,4 +1,5 @@
-from nbodykit.plugins import InputPainter, BoxSizeParser
+from nbodykit.plugins import InputPainter
+from nbodykit.utils.pluginargparse import BoxSizeParser
 
 import numpy
 import logging
